Remove `signal` function from Windows
This never did anything, despite being "used" in the Supervisor for
terminating a supervisor process.

Needless to say, this didn't actually end up terminating the
Supervisor (see https://github.com/habitat-sh/habitat/issues/4945).

Corresponding changes will need to be made in the Supervisor to not
call this code on Windows.

Signed-off-by: Christopher Maier <cmaier@chef.io>